(pc-23972) refacto: extract helper to shared/offer
